requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
pydantic==2.7.1
structlog==24.1.0
httpx==0.27.0
python-dotenv==1.0.1
